Add `extern_enums` arg to macros
The `extern_enums` arg is forwarded to the `graphql-client` crate, which
will use it as the type for the enum instead of generating code for it.

By default, the 3 large enums from Function APIs are passed as
`extern_enums`: LanguageCode, CountryCode, CurrencyCode. Users can
override that by sending `extern_enums` explicitly.